+++ src/agedi/cli/main.py
@@ -5,6 +5,7 @@ from agedi.cli.sample import sample
 from agedi.cli.inpaint import inpaint
 from agedi.cli.inspect import inspect
 from agedi.cli.predict import predict
+from agedi.cli.relax import relax
 
 @click.group()
 @click.version_option()
@@ -18,4 +19,5 @@ cli.add_command(sample)
 cli.add_command(inpaint)
 cli.add_command(inspect)
 cli.add_command(predict)
+cli.add_command(relax)